rename @types/common to ConditionFn.ts
