+++ src/@types/ConditionFn.ts
